Feature - arbitrary sized maps & wrapping
This commit adds initial support for arbitrarily sized maps and
enabling/disabling map wrapping. This is still very much in development
and as such, there are many artefacts on the map at this time. Many
portions of the code have also been rewritten to provide a more
consistent library that is easy to modify.