Cleaned up mapgen command and started to rename it.
Parts of the cli, used to be called mapGenCLI and other parts used to
refer to it as CLI_mapGen. Both these names feel quite clumsy and don't
really reflect the intention of the project properly, thus hence forth
the name of the terminal command will be called mgc (Short for map
generator cli).