Changed build system to CMAKE
Due to the projects age and size, this was a relatively trivial change.
This change was made due to CMAKE being a far more popular system, thus
most people will be more likely to have it installed on their system.
This also uncovered an issue with the way I had written the template
library, this is now also fixed.